Web description and fix delete tags

+++ backend/src/controllers/tag.controller.js
@@ -18,31 +18,34 @@ export const createTag = async (req, res) => {
 
 export const deleteTag = async (req, res) => {
     try {
-        const { key } = req.params;
-        const tag = await Tag.findOne({ key });
-        if (tag) {
-            const problems = await Problem.find({ 'tags.key': key });
+        if (req.isAdmin) {
+            const { key } = req.params;
+            const tag = await Tag.findOne({ key });
+            if (tag) {
+                const problems = await Problem.find({ 'tags.key': key });
 
-            if (problems.length != 0) {
-                const promises = [];
-                
-                problems.forEach(problem => {
-                    promises.push(Problem.findOneAndUpdate(
-                        { problemId: problem.problemId },
-                        { $pull: { tags: { key }}},
-                        {
-                            new: true
-                        }
-                    ));
-                });
+                if (problems.length != 0) {
+                    const promises = [];
+                    
+                    problems.forEach(problem => {
+                        promises.push(Problem.findOneAndUpdate(
+                            { problemId: problem.problemId },
+                            { $pull: { tags: { key } } },
+                            {
+                                new: true
+                            }
+                        ));
+                    });
 
-                await Promise.all(promises);
+                    await Promise.all(promises);
+                }
+
+                await Tag.findOneAndDelete({ key });
+                return res.status(200).json();
             }
-
-            await Tag.findOneAndDelete({ key });
-            return res.status(200).json();
+            return res.status(404).json({ message: "Tag not found" });
         }
-        return res.status(404).json({ message: "Tag not found" });
+        return res.status(401).json({ message: "Unauthorized" });
     } catch (error) {
         console.log(error);
         return res.status(500).json({ message: "Error" });
